requirements.txt
# AISG KKD Denetim Sistemi - Python Bağımlılıkları

# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0

# Computer Vision ve AI/ML
opencv-python==4.8.1.78
ultralytics==8.0.196
torch>=1.13.0
torchvision>=0.14.0

# Görüntü İşleme
Pillow==10.0.1
numpy==1.24.3

# Web ve API
requests==2.31.0
gunicorn==21.2.0

# Veri İşleme
pandas==2.0.3
matplotlib==3.7.2

# Güvenlik ve Kimlik Doğrulama
python-dotenv==1.0.0
Flask-Login==0.6.3

# Geliştirme Araçları
python-dateutil==2.8.2
